src/main/java/br/com/bookstoreapi/clients/client/service/SaveClientServiceImpl.java
package com.bookstoreapi.bookstoreapi.client.service;

import com.bookstoreapi.bookstoreapi.client.Client;
import com.bookstoreapi.bookstoreapi.client.ClientRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.UUID;

@RequiredArgsConstructor
@Service
public class SaveClientServiceImpl implements SaveClientService {

    private final ClientRepository clientRepository;


    @Override
    public Client save(Client client) {
        client.setUuid(UUID.randomUUID());
        return clientRepository.save(client);
    }
}
